+++ MetalfluxApi.Server/Core/Base/IService.cs
@@ -2,9 +2,9 @@
 
 public interface IService<TDto, TModel>
 {
-    TDto Get(int id);
+    TDto Get(long id);
     TDto Add(TDto item);
-    int Remove(int id);
+    long Remove(long id);
     TDto Update(TDto item);
     TDto ToDto(TModel model);
     public TModel ToModel(TDto dto);
@@ -12,10 +12,10 @@ public interface IService<TDto, TModel>
 
 public interface IAuditableService<TDto, TModel, in TCreateDto, in TUpdateDto>
 {
-    TDto Get(int id);
-    TDto Add(TCreateDto item, int createdById);
-    int Remove(int id);
-    TDto Update(TUpdateDto item, int updatedById);
+    TDto Get(long id);
+    TDto Add(TCreateDto item, long createdById);
+    long Remove(long id);
+    TDto Update(TUpdateDto item, long updatedById);
     TDto ToDto(TModel model);
     public TModel ToModelForCreation(TCreateDto dto);
     public TModel ToModelForUpdate(TUpdateDto dto);
