+++ MetalfluxApi.Server/Core/Dto/CursorResponse.cs
@@ -3,6 +3,6 @@
 public struct CursorResponse<TDto>
 {
     public List<TDto> Data { get; set; }
-    public int NextCursor { get; set; }
+    public long NextCursor { get; set; }
     public bool LastItemReached { get; set; }
 }
